Made money and timer scriptable objects
Made money and timer scriptable objects
