+++ Assets/Scenes/Combined scenes/Week 1 stuff.unity
@@ -1326,8 +1326,8 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   moneyText: {fileID: 24610674}
-  genPerSec: 2
-  moneyAmount: 0
+  genPerSec: {fileID: 11400000, guid: 9d0ae6bf0ab46402da3b14888a7daa84, type: 2}
+  moneyAmount: {fileID: 11400000, guid: 91aec1bcab1f941ac8249cb8d32e8614, type: 2}
 --- !u!1 &351274800
 GameObject:
   m_ObjectHideFlags: 0
@@ -2849,7 +2849,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   timerText: {fileID: 1293617363}
-  remainingTimeSeconds: 600
+  remainingTimeSeconds: {fileID: 11400000, guid: bab981189e2514c6cb5b0f4b61443103, type: 2}
 --- !u!1 &1216897151
 GameObject:
   m_ObjectHideFlags: 0
