+++ Assets/scripts/Game systems/FloatVariable.cs
@@ -1,0 +1,11 @@
+using JetBrains.Annotations;
+using UnityEngine;
+[CreateAssetMenu]
+public class FloatVariable : ScriptableObject
+{
+    [SerializeField] float _BaseValue = 600;
+    void OnEnable () { _Value =  _BaseValue;}
+    public float _Value; 
+
+    
+}
