+++ Assets/scripts/Game systems/FloatVariable.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: f8ed8d2c015394dd38fa765e1776b8b2
