+++ Assets/scripts/Game systems/Money.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f5dcc7c32713249ce9a35f5f0c042b83
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
