+++ Assets/scripts/Game systems/Money/Money.cs.meta
