+++ Assets/scripts/Game systems/Money/MoneyAmount.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 91aec1bcab1f941ac8249cb8d32e8614
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
