+++ Assets/scripts/Game systems/Money/MoneyPerSec.asset
@@ -1,0 +1,16 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f8ed8d2c015394dd38fa765e1776b8b2, type: 3}
+  m_Name: MoneyPerSec
+  m_EditorClassIdentifier: 
+  _BaseValue: 2
+  _Value: 2
