+++ Assets/scripts/Game systems/Money/MoneyPerSec.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9d0ae6bf0ab46402da3b14888a7daa84
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
