+++ Assets/scripts/Game systems/Time.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aad94de117c394f398dfdddd2a239df5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
