+++ Assets/scripts/Game systems/Time/TimeRemaining.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bab981189e2514c6cb5b0f4b61443103
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
