+++ Assets/scripts/Game systems/Time/Timer.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 3fcd4ad2cd3c04e6ca857e9eedb2a234
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences:
+  - timerText: {instanceID: 0}
+  - remainingTimeSeconds: {fileID: 11400000, guid: bab981189e2514c6cb5b0f4b61443103, type: 2}
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
