--- Assets/scripts/Game systems/Timer.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 3fcd4ad2cd3c04e6ca857e9eedb2a234
